source_code_agent/requirements.txt
# Web框架
fastapi
uvicorn[standard]
pydantic
python-multipart
starlette
httpx

# 数据库
sqlalchemy
alembic
psycopg2-binary
pymysql

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt
pyyaml
python-dotenv==1.0.0
email-validator

# 文件处理
modelscope
watchdog
cachetools
anthropic
magic_pdf

# 异步支持
aiohttp
aiofiles==23.2.1

# 日志
loguru

# 向量数据库支持
faiss-cpu
sentence-transformers
scikit-learn
numpy
torch

# 知识库向量化支持
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# 向量库
pymilvus

# 新增依赖
pillow==10.0.0
pandas
openpyxl==3.1.2
python-docx==1.0.1
pypdf==3.15.5
neo4j
neo4j-graphrag==1.3.0
minio>=7.2.0

asyncpg
aiomysql

tavily-python

# MCP服务支持
fastmcp>=0.1.0
mysql-connector-python
google-genai
google-generativeai
mineru[core]